게시글 수정, 삭제 API 추가 및 Impl postfix 삭제
